userInteractionService/requirements.txt
Django==3.1.3
djangorestframework==3.12.2 
psycopg2==2.8.6 #try psycopg2-binary if it doesn't work
pandas==1.4.0
pika==1.2.0
requests==2.27